+++ tests/chrome_os.php
@@ -1,0 +1,42 @@
+<?php
+
+error_reporting(E_ALL);
+
+require_once('../src/BrowserDetection.php');
+$Browser = new foroco\BrowserDetection();
+
+// Testing new Chrome OS / Chromium OS User-Agents
+
+$useragent[] = 'Mozilla/5.0 (X11; CrOS x86_64 14388.61.0) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.147 Safari/537.36';
+$useragent[] = 'Mozilla/5.0 (X11; CrOS aarch64 14388.61.0) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/99.0.4844.57 Safari/537.36';
+$useragent[] = 'Mozilla/5.0 (X11; CrOS armv7l 14388.61.0) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/99.0.4844.57 Safari/537.36';
+$useragent[] = 'Mozilla/5.0 (X11; Chromium OS x86_64 14388.61.0) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/99.0.4844.57 Safari/537.36';
+
+$time_start = microtime(true);
+
+foreach ($useragent as $u)
+{
+
+	// Detect all possible environment data from User-Agents
+	$result = $Browser->getAll($u);
+
+	echo '<div style="font-size:16px; font-weight:bold"><pre>';
+	echo $u;
+	echo '</pre></div>';
+
+	echo '<div style="font-size:18px; font-weight:bold"><pre>';
+	print_r($result);
+	echo '</pre></div>';
+
+	echo '<div style="font-size:16px; font-weight:bold"><pre>';
+	echo '--------------------------------------------------------------------------------';
+	echo '</pre></div>';
+
+}
+
+$time_end = microtime(true);
+$time_result = $time_end - $time_start;
+
+echo '<p style="font-size:21px">Total execution time: '.substr($time_result,0,7).' sec.</p>';
+
+?>