+++ tests/float_issue.php
@@ -1,0 +1,41 @@
+<?php
+
+error_reporting(E_ALL);
+
+require_once('../src/BrowserDetection.php');
+$Browser = new foroco\BrowserDetection();
+
+// Float issue
+
+$useragent[] = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_6) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/13.10.2 Safari/605.1.15';
+$useragent[] = 'Mozilla/5.0 (Windows NT 6.2; WOW64) AppleWebKit/537.10 (KHTML, like Gecko) lswebkit Safari/537.10  Safari/537.10';
+$useragent[] = 'Mozilla/5.0 (iPad; CPU OS 11_0_3 like Mac OS X) AppleWebKit/604.1.38 (KHTML, like Gecko) OPiOS/16.10.3.120766 Mobile/15A432 Safari/9537.53';
+
+$time_start = microtime(true);
+
+foreach ($useragent as $u)
+{
+
+	// Detect all possible environment data from User-Agents
+	$result = $Browser->getAll($u);
+
+	echo '<div style="font-size:16px; font-weight:bold"><pre>';
+	echo $u;
+	echo '</pre></div>';
+
+	echo '<div style="font-size:18px; font-weight:bold"><pre>';
+	print_r($result);
+	echo '</pre></div>';
+
+	echo '<div style="font-size:16px; font-weight:bold"><pre>';
+	echo '--------------------------------------------------------------------------------';
+	echo '</pre></div>';
+
+}
+
+$time_end = microtime(true);
+$time_result = $time_end - $time_start;
+
+echo '<p style="font-size:21px">Total execution time: '.substr($time_result,0,7).' sec.</p>';
+
+?>