+++ tests/ios_webview.php
@@ -7,6 +7,7 @@ $Browser = new foroco\BrowserDetection();
 
 // Testing some new User-Agents
 
+$useragent[] = 'Mozilla/5.0 (iPhone; CPU iPhone OS 13_5 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Mobile/15E148';
 $useragent[] = 'Mozilla/5.0 (iPad; CPU OS 10_3_4 like Mac OS X) AppleWebKit/600.1.4 (KHTML, like Gecko) GSA/9.0.60246 Mobile/14G61 Safari/600.1.4';
 $useragent[] = 'Mozilla/5.0 (iPad; CPU OS 13_3 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Mobile/15E148 OPT/2.5.3';
 $useragent[] = 'Mozilla/5.0 (iPhone; CPU iPhone OS 14_1 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Mobile/15E148 SRCHAPP';
@@ -20,6 +21,14 @@ $useragent[] = 'Mozilla/5.0 (iPhone; CPU iPhone OS 13_7 like Mac OS X) AppleWebK
 $useragent[] = 'Mozilla/5.0 (iPhone; CPU iPhone OS 13_7 like Mac OS X; ru-RU) AppleWebKit/537.36 (KHTML, like Gecko) Version/13.7 Mobile/17H35 Safari/537.36 Puffin/5.2.4IP';
 $useragent[] = 'Mozilla/5.0 (iPhone; CPU iPhone OS 11_3_1 like Mac OS X) AppleWebKit/605.1.33 (KHTML, like Gecko) Mobile/15E302 YaBrowser/19.5.2.38.10 YaApp_iOS/25.00 YaApp_iOS_Browser/25.00 Safari/604.1';
 $useragent[] = 'Mozilla/5.0 (iPhone; CPU iPhone OS 12_4_1 like Mac OS X) AppleWebKit/601.1.46 (KHTML, like Gecko) Version/12.0 YaBrowser/16.4.1.9336.10 Mobile/16G102 Safari/601.1';
+$useragent[] = 'Mozilla/5.0 (iPhone; CPU iPhone OS 16_2 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Mobile/15E148 YaBrowser/23.7.0.2091.10 YaApp_iOS/2307.0 YaApp_iOS_Browser/2307.0 Safari/604.1 SA/3';
+$useragent[] = 'Mozilla/5.0 (iPhone; CPU iPhone OS 16_5_1 like Mac OS X; ru) AppleWebKit/537.51.1 (KHTML, like Gecko) Mobile/20F75 UCBrowser/11.3.5.1203 Mobile';
+$useragent[] = 'Mozilla/5.0 (iPhone; CPU iPhone OS 16_2 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko)  Mobile/15E148 Safari/604.1';
+$useragent[] = 'Mozilla/5.0 (iPhone; CPU iPhone OS 16_6 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Mobile/15E148 Safari/605.1.15';
+$useragent[] = 'Mozilla/5.0 (iPhone; CPU iPhone OS 17_1_2 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Mobile/15E148 SP-engine/2.85.0 main/1.0 baiduboxapp/13.47.0.10 (Baidu; P2 17.1.2) NABar/1.0 themeUA=Theme/default';
+$useragent[] = 'Mozilla/5.0 (iPhone; CPU iPhone OS 10_0_1 like Mac OS X) AppleWebKit/602.1.32 (KHTML, like Gecko) Mobile/14A403 Twitter for iPhone';
+$useragent[] = 'Mozilla/5.0 (iPhone; CPU iPhone OS 10_1_1 like Mac OS X) AppleWebKit/602.2.14 (KHTML, like Gecko) Mobile/14B100 KAKAOTALK 5.9.2';
+$useragent[] = 'Mozilla/5.0 (iPhone; CPU iPhone OS 10_1_1 like Mac OS X) AppleWebKit/602.2.14 (KHTML, like Gecko) Mobile/14B100 MicroMessenger/6.3.30 NetType/WIFI Language/en';
 
 $time_start = microtime(true);